[Feature #161209043] Create views folder
