+++ requirements.txt
@@ -6,6 +6,7 @@ itsdangerous==0.24
 Jinja2==2.10
 MarkupSafe==1.0
 nose==1.3.7
+python-dotenv==0.9.1
 pytz==2018.5
 six==1.11.0
 Werkzeug==0.14.1
